frontend/requirements.txt
streamlit==1.38.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
langchain==0.2.6
langchain-text-splitters==0.2.2
boto3==1.34.162
lancedb==0.13.0
rerankers[transformers]==0.5.3
fastembed==0.3.6
tantivy==0.22.0

--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.3.1+cpu
